tidy, move configuration and remove oninstall
