+++ server/plugin.go
@@ -1,12 +1,10 @@
 package main
 
 import (
-	"reflect"
 	"sync"
 	"time"
 
 	"github.com/mattermost/mattermost-plugin-starter-template/server/command"
-	"github.com/mattermost/mattermost-plugin-starter-template/server/config"
 	"github.com/mattermost/mattermost-plugin-starter-template/server/store/kvstore"
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/plugin"
@@ -35,11 +33,7 @@ type Plugin struct {
 
 	// configuration is the active plugin configuration. Consult getConfiguration and
 	// setConfiguration for usage.
-	configuration *config.Configuration
-}
-
-func (p *Plugin) OnInstall() error {
-	return nil
+	configuration *configuration
 }
 
 // OnActivate is invoked when the plugin is activated. If an error is returned, the plugin will be deactivated.
@@ -78,61 +72,6 @@ func (p *Plugin) OnDeactivate() error {
 // This will execute the commands that were registered in the NewCommandHandler function.
 func (p *Plugin) ExecuteCommand(c *plugin.Context, args *model.CommandArgs) (*model.CommandResponse, *model.AppError) {
 	return p.commandClient.Handle(args)
-}
-
-// getConfiguration retrieves the active configuration under lock, making it safe to use
-// concurrently. The active configuration may change underneath the client of this method, but
-// the struct returned by this API call is considered immutable.
-// func (p *Plugin) getConfiguration() *config.Configuration {
-// 	p.configurationLock.RLock()
-// 	defer p.configurationLock.RUnlock()
-
-// 	if p.configuration == nil {
-// 		return &config.Configuration{}
-// 	}
-
-// 	return p.configuration
-// }
-
-// setConfiguration replaces the active configuration under lock.
-//
-// Do not call setConfiguration while holding the configurationLock, as sync.Mutex is not
-// reentrant. In particular, avoid using the plugin API entirely, as this may in turn trigger a
-// hook back into the plugin. If that hook attempts to acquire this lock, a deadlock may occur.
-//
-// This method panics if setConfiguration is called with the existing configuration. This almost
-// certainly means that the configuration was modified without being cloned and may result in
-// an unsafe access.
-func (p *Plugin) setConfiguration(configuration *config.Configuration) {
-	p.configurationLock.Lock()
-	defer p.configurationLock.Unlock()
-
-	if configuration != nil && p.configuration == configuration {
-		// Ignore assignment if the configuration struct is empty. Go will optimize the
-		// allocation for same to point at the same memory address, breaking the check
-		// above.
-		if reflect.ValueOf(*configuration).NumField() == 0 {
-			return
-		}
-
-		panic("setConfiguration called with the existing configuration")
-	}
-
-	p.configuration = configuration
-}
-
-// OnConfigurationChange is invoked when configuration changes may have been made. It is called once before OnActivate.
-func (p *Plugin) OnConfigurationChange() error {
-	var configuration = new(config.Configuration)
-
-	// Load the public configuration fields from the Mattermost server configuration.
-	if err := p.API.LoadPluginConfiguration(configuration); err != nil {
-		return errors.Wrap(err, "failed to load plugin configuration")
-	}
-
-	p.setConfiguration(configuration)
-
-	return nil
 }
 
 // See https://developers.mattermost.com/extend/plugins/server/reference/
